Replace the silver searcher with ripgrep
The main advantage of ripgrep is its proper support for .gitignore files.
There are many bugs in the silver searcher around this feature, but
ripgrep seems to deal even with somewhat less common .gitignore rules.

Besides, ripgrep is also faster than the silver searcher. The latter is
already really fast, but with ripgrep the initial lag is gone – the CtrlP
window opens immediately after pressing the shortcut.

The transition would be much harder if not for Otis Wright post about
using ripgrep with vim and fzf.

https://medium.com/@crashybang/supercharge-vim-with-fzf-and-ripgrep-d4661fc853d2