alabaster: Go back to using pure black for regular text
Initially I changed it to #222222, following the common wisdom of not showing
pure black on white background. However, we don't use a white background
in this colorscheme. Instead we use #DFDFDF.

Using WebAIM contrast checker, the results look as follows:

Text    |Background |Contrast ratio
:------:|:---------:|:------------:
#000000 |#FFFFFF    |21:1
#222222 |#FFFFFF    |15.9:1
#000000 |#DFDFDF    |15.75:1
#222222 |#DFDFDF    |11.91:1

So our previous colors (#000 for text and #DFDFDF for background) already
had more or less the same contrast ratio as #222 on #FFF.